rename shortcut script generation project
i felt the existing name did not reflect the contents of the project well enough, and it also did not follow the conventions of other project names in the solution.
