--- source/AutoHotkeyScripts/Enums/ShortcutScriptEntryLineTypes.cs
@@ -1,12 +1,0 @@
-﻿namespace Petrichor.AutoHotkeyScripts.Enums
-{
-	public enum ShortcutScriptEntryLineTypes
-	{
-		Unknown,
-		EntryStart,
-		EntryEnd,
-		Identity,
-		Pronoun,
-		Decoration,
-	}
-}
