--- source/AutoHotkeyScripts/Exceptions/BlankInputFieldException.cs
@@ -1,10 +1,0 @@
-﻿namespace Petrichor.AutoHotkeyScripts.Exceptions
-{
-	[ Serializable ]
-	public class BlankInputFieldException : Exception
-	{
-		public BlankInputFieldException() : base() { }
-		public BlankInputFieldException( string message ) : base( message ) { }
-		public BlankInputFieldException( string message, Exception inner ) : base( message, inner ) { }
-	}
-}
