--- source/AutoHotkeyScripts/Exceptions/DuplicateInputFieldException.cs
@@ -1,10 +1,0 @@
-﻿namespace Petrichor.AutoHotkeyScripts.Exceptions
-{
-	[ Serializable ]
-	public class DuplicateInputFieldException : Exception
-	{
-		public DuplicateInputFieldException() : base() { }
-		public DuplicateInputFieldException( string message ) : base( message ) { }
-		public DuplicateInputFieldException( string message, Exception inner ) : base( message, inner ) { }
-	}
-}
