--- source/AutoHotkeyScripts/Exceptions/InputEntryNotClosedException.cs
@@ -1,10 +1,0 @@
-﻿namespace Petrichor.AutoHotkeyScripts.Exceptions
-{
-	[ Serializable ]
-	public class InputEntryNotClosedException : Exception
-	{
-		public InputEntryNotClosedException() : base() { }
-		public InputEntryNotClosedException( string message ) : base( message ) { }
-		public InputEntryNotClosedException( string message, Exception inner ) : base( message, inner ) { }
-	}
-}
