--- source/AutoHotkeyScripts/Exceptions/InvalidInputFieldException.cs
@@ -1,10 +1,0 @@
-﻿namespace Petrichor.AutoHotkeyScripts.Exceptions
-{
-	[ Serializable ]
-	public class InvalidInputFieldException : Exception
-	{
-		public InvalidInputFieldException() : base() { }
-		public InvalidInputFieldException( string message ) : base( message ) { }
-		public InvalidInputFieldException( string message, Exception inner ) : base( message, inner ) { }
-	}
-}
