--- source/AutoHotkeyScripts/Exceptions/MissingInputFieldException.cs
@@ -1,10 +1,0 @@
-﻿namespace Petrichor.AutoHotkeyScripts.Exceptions
-{
-	[ Serializable ]
-	public class MissingInputFieldException : Exception
-	{
-		public MissingInputFieldException() : base() { }
-		public MissingInputFieldException( string message ) : base( message ) { }
-		public MissingInputFieldException( string message, Exception inner ) : base( message, inner ) { }
-	}
-}
