--- source/AutoHotkeyScripts/Exceptions/UnexpectedCharacterException.cs
@@ -1,10 +1,0 @@
-﻿namespace Petrichor.AutoHotkeyScripts.Exceptions
-{
-	[ Serializable ]
-	public class UnexpectedCharacterException : Exception
-	{
-		public UnexpectedCharacterException() : base() { }
-		public UnexpectedCharacterException( string message ) : base( message ) { }
-		public UnexpectedCharacterException( string message, Exception inner ) : base( message, inner ) { }
-	}
-}
