--- source/AutoHotkeyScripts/Utilities/ShortcutScriptMetadataParser.cs
@@ -1,17 +1,0 @@
-﻿using Petrichor.AutoHotkeyScripts.Containers;
-
-
-namespace Petrichor.AutoHotkeyScripts.Utilities
-{
-	public class ShortcutScriptMetadataParser
-	{
-		private const string DefaultIconFilePathToken = "defaultIcon:";
-		private StringTokenParser TokenParser = new StringTokenParser();
-
-
-		public ShortcutScriptMetadata ParseMetadataFromData( string[] data, ref int i )
-		{
-			throw new NotImplementedException();
-		}
-	}
-}
