+++ source/ShortcutScriptGeneration/Enums/ShortcutScriptEntryLineTypes.cs
@@ -1,0 +1,12 @@
+﻿namespace Petrichor.ShortcutScriptGeneration.Enums
+{
+	public enum ShortcutScriptEntryLineTypes
+	{
+		Unknown,
+		EntryStart,
+		EntryEnd,
+		Identity,
+		Pronoun,
+		Decoration,
+	}
+}
