+++ source/ShortcutScriptGeneration/Exceptions/BlankInputFieldException.cs
@@ -1,0 +1,10 @@
+﻿namespace Petrichor.ShortcutScriptGeneration.Exceptions
+{
+	[ Serializable ]
+	public class BlankInputFieldException : Exception
+	{
+		public BlankInputFieldException() : base() { }
+		public BlankInputFieldException( string message ) : base( message ) { }
+		public BlankInputFieldException( string message, Exception inner ) : base( message, inner ) { }
+	}
+}
