+++ source/ShortcutScriptGeneration/Exceptions/EscapeCharacterMismatchException.cs
@@ -1,0 +1,10 @@
+﻿namespace Petrichor.ShortcutScriptGeneration.Exceptions
+{
+	[ Serializable ]
+	public class EscapeCharacterMismatchException : Exception
+	{
+		public EscapeCharacterMismatchException() : base() { }
+		public EscapeCharacterMismatchException( string message ) : base( message ) { }
+		public EscapeCharacterMismatchException( string message, Exception inner ) : base( message, inner ) { }
+	}
+}
