+++ source/ShortcutScriptGeneration/Exceptions/InputEntryNotClosedException.cs
@@ -1,0 +1,10 @@
+﻿namespace Petrichor.ShortcutScriptGeneration.Exceptions
+{
+	[ Serializable ]
+	public class InputEntryNotClosedException : Exception
+	{
+		public InputEntryNotClosedException() : base() { }
+		public InputEntryNotClosedException( string message ) : base( message ) { }
+		public InputEntryNotClosedException( string message, Exception inner ) : base( message, inner ) { }
+	}
+}
