+++ source/ShortcutScriptGeneration/Exceptions/InvalidInputFieldException.cs
@@ -1,0 +1,10 @@
+﻿namespace Petrichor.ShortcutScriptGeneration.Exceptions
+{
+	[ Serializable ]
+	public class InvalidInputFieldException : Exception
+	{
+		public InvalidInputFieldException() : base() { }
+		public InvalidInputFieldException( string message ) : base( message ) { }
+		public InvalidInputFieldException( string message, Exception inner ) : base( message, inner ) { }
+	}
+}
