+++ source/ShortcutScriptGeneration/Exceptions/MissingInputFieldException.cs
@@ -1,0 +1,10 @@
+﻿namespace Petrichor.ShortcutScriptGeneration.Exceptions
+{
+	[ Serializable ]
+	public class MissingInputFieldException : Exception
+	{
+		public MissingInputFieldException() : base() { }
+		public MissingInputFieldException( string message ) : base( message ) { }
+		public MissingInputFieldException( string message, Exception inner ) : base( message, inner ) { }
+	}
+}
