+++ source/ShortcutScriptGeneration/Exceptions/RegionNotClosedException.cs
@@ -1,0 +1,10 @@
+﻿namespace Petrichor.ShortcutScriptGeneration.Exceptions
+{
+	[ Serializable ]
+	public class RegionNotClosedException : Exception
+	{
+		public RegionNotClosedException() : base() { }
+		public RegionNotClosedException( string message ) : base( message ) { }
+		public RegionNotClosedException( string message, Exception inner ) : base( message, inner ) { }
+	}
+}
