+++ source/ShortcutScriptGeneration/Exceptions/UnexpectedCharacterException.cs
@@ -1,0 +1,10 @@
+﻿namespace Petrichor.ShortcutScriptGeneration.Exceptions
+{
+	[ Serializable ]
+	public class UnexpectedCharacterException : Exception
+	{
+		public UnexpectedCharacterException() : base() { }
+		public UnexpectedCharacterException( string message ) : base( message ) { }
+		public UnexpectedCharacterException( string message, Exception inner ) : base( message, inner ) { }
+	}
+}
