+++ source/ShortcutScriptGeneration/Exceptions/UnknownTokenException.cs
@@ -1,0 +1,10 @@
+﻿namespace Petrichor.ShortcutScriptGeneration.Exceptions
+{
+	[ Serializable ]
+	public class UnknownTokenException : Exception
+	{
+		public UnknownTokenException() : base() { }
+		public UnknownTokenException( string message ) : base( message ) { }
+		public UnknownTokenException( string message, Exception inner ) : base( message, inner ) { }
+	}
+}
