+++ source/ShortcutScriptGeneration/LookUpTables/ShortcutScriptTemplateMarkers.cs
@@ -1,0 +1,25 @@
+﻿namespace Petrichor.ShortcutScriptGeneration.LookUpTables
+{
+	public static class ShortcutScriptTemplateMarkers
+	{
+		public static readonly Dictionary< char, string > LookUpTable = new Dictionary< char, string >()
+		{
+			{
+				'#',
+				"name"
+			},
+			{
+				'@',
+				"tag"
+			},
+			{
+				'$',
+				"pronoun"
+			},
+			{
+				'&',
+				"decoration"
+			},
+		};
+	}
+}
