+++ source/ShortcutScriptGeneration/ShortcutScriptGeneration.csproj
