+++ source/ShortcutScriptGeneration/Utilities/ShortcutScriptMetadataParser.cs
@@ -1,0 +1,17 @@
+﻿using Petrichor.ShortcutScriptGeneration.Containers;
+
+
+namespace Petrichor.ShortcutScriptGeneration.Utilities
+{
+	public class ShortcutScriptMetadataParser
+	{
+		private const string DefaultIconFilePathToken = "defaultIcon:";
+		private StringTokenParser TokenParser = new StringTokenParser();
+
+
+		public ShortcutScriptMetadata ParseMetadataFromData( string[] data, ref int i )
+		{
+			throw new NotImplementedException();
+		}
+	}
+}
