+++ source/ShortcutScriptGenerationTests/Utilities/ShortcutScriptGeneratorTests.cs
@@ -1,0 +1,88 @@
+﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
+using Petrichor.ShortcutScriptGeneration.Containers;
+using Petrichor.Common.Info;
+using Petrichor.Logging;
+using Petrichor.TestShared.Info;
+using Petrichor.TestShared.Utilities;
+
+
+namespace Petrichor.ShortcutScriptGeneration.Utilities.Tests
+{
+	[ TestClass ]
+	public class ShortcutScriptGeneratorTests
+	{
+		public static class TestData
+		{
+			public static readonly ShortcutScriptEntry[] Entries = new ShortcutScriptEntry[]
+			{
+				new ShortcutScriptEntry( new List<ShortcutScriptIdentity>(){ new ShortcutScriptIdentity( "name", "tag" ) }, "pronoun", "decoration" ),
+			};
+			public static readonly string[] Templates = new string[]
+			{
+				"::@`tag`:: `name`",
+				"::@$&`tag`:: `name` `pronoun` `decoration`",
+			};
+			public static readonly ShortcutScriptInput Input = new ShortcutScriptInput( Entries, Templates );
+			public static readonly string[] Macros = new string[]
+			{
+				"::@tag:: name",
+				"::@$&tag:: name pronoun decoration",
+			};
+			public static readonly string[] GeneratedOutputFileContents = new string[]
+			{
+				$"; Generated by { AppInfo.AppNameAndVersion } AutoHotkey shortcut script generator",
+				"; https://github.com/SparkliTtwizzl/petrichor",
+				"",
+				"",
+				"#SingleInstance Force",
+				"",
+				"::@tag:: name",
+				"::@$&tag:: name pronoun decoration",
+			};
+		}
+
+
+		public ShortcutScriptGenerator? ScriptGenerator;
+
+
+		[TestInitialize ]
+		public void Setup()
+		{
+			TestUtilities.InitializeLoggingForTests();
+			ScriptGenerator = new ShortcutScriptGenerator();
+		}
+
+
+		[ TestMethod ]
+		public void GenerateMacrosFromTemplatesTest_Success()
+		{
+			var expected = TestData.Macros;
+			var actual = ScriptGenerator.GenerateMacrosFromInput( TestData.Input ).ToArray();
+
+			Log.WriteLine( "expected:" );
+			foreach ( var line in expected )
+			{
+				Log.WriteLine( $"[{ line }]" );
+			}
+			Log.WriteLine();
+			Log.WriteLine( "actual:" );
+			foreach ( var line in actual )
+			{
+				Log.WriteLine( $"[{ line }]" );
+			}
+
+			CollectionAssert.AreEqual( expected, actual );
+		}
+
+		[ TestMethod ]
+		public void GenerateScriptTest_Success()
+		{
+			var outputFile = $@"{ TestDirectories.TestOutputDirectory }\{ nameof( ShortcutScriptGenerator ) }_{ nameof( GenerateScriptTest_Success ) }.ahk";
+			ScriptGenerator.GenerateScript( TestData.Macros, outputFile );
+
+			var expected = TestData.GeneratedOutputFileContents;
+			var actual = File.ReadAllLines( outputFile );
+			CollectionAssert.AreEqual( expected, actual );
+		}
+	}
+}
